refactor(posts): split postApi, fix TS 'any' in updateQueryData, and improve optimistic comment caching
