+++ src/entities/posts/api/postApi.ts
@@ -1,11 +1,7 @@
 import {
   CreatePostInputDto,
-  GetCommentAnswersParams,
-  GetCommentsParams,
   GetPostsByUserParams,
   GetPostsParams,
-  PaginatedAnswersResponse,
-  PaginatedCommentsResponse,
   PaginatedPosts,
   PaginatedResponse,
   PostImageViewModel,
@@ -13,20 +9,7 @@ import {
   UpdateLikeStatusDto,
   UpdatePostInputDto,
 } from '@/entities/posts/api/posts.types'
-import {
-  COMMENTS_PAGE_SIZE,
-  createOptimisticId,
-  getAnswersNextPageParam,
-  getCommentsNextPageParam,
-  incrementCommentAnswerCount,
-  patchAnswerLikeInPages,
-  patchCommentLikeInPages,
-  patchPostLikeFields,
-  prependAnswerToPages,
-  prependCommentToPages,
-  replaceAnswerInPages,
-  replaceCommentInPages,
-} from '@/entities/posts/lib/comment-likes'
+import { patchPostLikeFields } from '@/entities/posts/lib/comment-likes'
 import {
   getCreatePostInvalidationTags,
   getPostMutationInvalidationTags,
@@ -41,13 +24,9 @@ import {
 import { API_ROUTES } from '@/shared/api'
 import { baseApi } from '@/shared/api/base-api'
 import { showToastAlert } from '@/shared/lib'
-import { AnswersViewModel, CommentsViewModel, CreateCommentDto } from '@/shared/types/comments'
-import { UserBase } from '@/shared/types/user/models'
-import { InfiniteData } from '@reduxjs/toolkit/query'
 
 export const postApi = baseApi.injectEndpoints({
   endpoints: builder => ({
-    // ===== POSTS =====
     createPost: builder.mutation<PostViewModel, { body: CreatePostInputDto; userId: number }>({
       query: ({ body }) => ({
         url: API_ROUTES.POSTS.BASE,
@@ -275,266 +254,6 @@ export const postApi = baseApi.injectEndpoints({
         }
       },
     }),
-
-    // ===== COMMENTS =====
-    getPostComments: builder.infiniteQuery<PaginatedCommentsResponse, GetCommentsParams, number>({
-      infiniteQueryOptions: {
-        initialPageParam: 1,
-        getNextPageParam: getCommentsNextPageParam,
-      },
-      query: ({ pageParam, queryArg }) => ({
-        url: API_ROUTES.POSTS.COMMENTS(queryArg.postId),
-        params: {
-          pageSize: queryArg.pageSize ?? COMMENTS_PAGE_SIZE,
-          pageNumber: pageParam,
-          sortDirection: queryArg.sortDirection ?? 'desc',
-          sortBy: queryArg.sortBy,
-        },
-      }),
-      providesTags: (result, error, { postId }) => [{ type: 'Comments', id: postId }],
-      serializeQueryArgs: ({ endpointName, queryArgs }) => `${endpointName}-${queryArgs.postId}`,
-    }),
-
-    getCommentAnswers: builder.infiniteQuery<
-      PaginatedAnswersResponse,
-      GetCommentAnswersParams,
-      number
-    >({
-      infiniteQueryOptions: {
-        initialPageParam: 1,
-        getNextPageParam: getAnswersNextPageParam,
-      },
-      query: ({ pageParam, queryArg }) => ({
-        url: API_ROUTES.POSTS.COMMENT_ANSWERS(queryArg.postId, queryArg.commentId),
-        params: {
-          pageSize: queryArg.pageSize ?? COMMENTS_PAGE_SIZE,
-          pageNumber: pageParam,
-          sortDirection: queryArg.sortDirection ?? 'desc',
-          sortBy: queryArg.sortBy,
-        },
-      }),
-      providesTags: (result, error, { postId, commentId }) => [
-        { type: 'Comments', id: `${postId}-${commentId}` },
-      ],
-      serializeQueryArgs: ({ endpointName, queryArgs }) =>
-        `${endpointName}-${queryArgs.postId}-${queryArgs.commentId}`,
-    }),
-
-    updateCommentLikeStatus: builder.mutation<
-      void,
-      { postId: number; commentId: number; data: UpdateLikeStatusDto }
-    >({
-      query: ({ postId, commentId, data }) => ({
-        url: API_ROUTES.POSTS.LIKE_STATUS_COMMENT(postId, commentId),
-        method: 'PUT',
-        body: data,
-      }),
-      async onQueryStarted({ postId, commentId, data }, { dispatch, queryFulfilled }) {
-        const patchResult = dispatch(
-          postApi.util.updateQueryData(
-            'getPostComments',
-            { postId },
-            (draft: InfiniteData<PaginatedCommentsResponse, number>) => {
-              patchCommentLikeInPages(draft, commentId, data.likeStatus)
-            }
-          )
-        )
-
-        try {
-          await queryFulfilled
-        } catch {
-          patchResult.undo()
-          showToastAlert({ message: 'Failed to update like', type: 'error' })
-        }
-      },
-    }),
-
-    updateAnswerLikeStatus: builder.mutation<
-      void,
-      { postId: number; commentId: number; answerId: number; data: UpdateLikeStatusDto }
-    >({
-      query: ({ postId, commentId, answerId, data }) => ({
-        url: API_ROUTES.POSTS.LIKE_STATUS_ANSWER(postId, commentId, answerId),
-        method: 'PUT',
-        body: data,
-      }),
-      async onQueryStarted({ postId, commentId, answerId, data }, { dispatch, queryFulfilled }) {
-        const patchResult = dispatch(
-          postApi.util.updateQueryData(
-            'getCommentAnswers',
-            { postId, commentId },
-            (draft: InfiniteData<PaginatedAnswersResponse, number>) => {
-              patchAnswerLikeInPages(draft, answerId, data.likeStatus)
-            }
-          )
-        )
-
-        try {
-          await queryFulfilled
-        } catch {
-          patchResult.undo()
-          showToastAlert({ message: 'Failed to update like', type: 'error' })
-        }
-      },
-    }),
-
-    createComment: builder.mutation<
-      CommentsViewModel,
-      { postId: number; body: CreateCommentDto; optimisticFrom?: UserBase }
-    >({
-      query: ({ postId, body }) => ({
-        url: API_ROUTES.POSTS.CREATE_COMMENT(postId),
-        method: 'POST',
-        body,
-      }),
-      async onQueryStarted({ postId, body, optimisticFrom }, { dispatch, queryFulfilled }) {
-        const tempId = createOptimisticId()
-        const optimisticComment: CommentsViewModel = {
-          id: tempId,
-          postId,
-          content: body.content,
-          createdAt: new Date().toISOString(),
-          likeCount: 0,
-          isLiked: false,
-          answerCount: 0,
-          from: optimisticFrom ?? { id: 0, userName: 'You', avatars: [] },
-        }
-
-        const patchResult = dispatch(
-          postApi.util.updateQueryData(
-            'getPostComments',
-            { postId, pageSize: COMMENTS_PAGE_SIZE },
-            (draft: InfiniteData<PaginatedCommentsResponse, number>) => {
-              prependCommentToPages(draft, optimisticComment)
-            }
-          )
-        )
-
-        try {
-          const { data: createdComment } = await queryFulfilled
-
-          dispatch(
-            postApi.util.updateQueryData(
-              'getPostComments',
-              { postId, pageSize: COMMENTS_PAGE_SIZE },
-              (draft: InfiniteData<PaginatedCommentsResponse, number>) => {
-                replaceCommentInPages(draft, tempId, createdComment)
-              }
-            )
-          )
-        } catch {
-          patchResult.undo()
-          showToastAlert({ message: 'Failed to publish comment', type: 'error' })
-        }
-      },
-    }),
-
-    createAnswer: builder.mutation<
-      AnswersViewModel,
-      { postId: number; commentId: number; body: CreateCommentDto; optimisticFrom?: UserBase }
-    >({
-      query: ({ postId, commentId, body }) => ({
-        url: API_ROUTES.POSTS.CREATE_ANSWER_COMMENT(postId, commentId),
-        method: 'POST',
-        body,
-      }),
-      async onQueryStarted(
-        { postId, commentId, body, optimisticFrom },
-        { dispatch, queryFulfilled, getState }
-      ) {
-        const answersArgs = { postId, commentId, pageSize: COMMENTS_PAGE_SIZE }
-        const tempId = createOptimisticId()
-        const optimisticAnswer: AnswersViewModel = {
-          id: tempId,
-          commentId,
-          content: body.content,
-          createdAt: new Date().toISOString(),
-          likeCount: 0,
-          isLiked: false,
-          from: optimisticFrom ?? { id: 0, userName: 'You', avatars: [] },
-        }
-
-        const cacheEntry = postApi.endpoints.getCommentAnswers.select(answersArgs)(getState())
-        const hasCache = Boolean(cacheEntry.data?.pages?.length)
-
-        let undoAnswer: (() => void) | undefined
-
-        if (hasCache) {
-          const patchResult = dispatch(
-            postApi.util.updateQueryData(
-              'getCommentAnswers',
-              answersArgs,
-              (draft: InfiniteData<PaginatedAnswersResponse, number>) => {
-                prependAnswerToPages(draft, optimisticAnswer)
-              }
-            )
-          )
-
-          undoAnswer = () => patchResult.undo()
-        } else {
-          dispatch(
-            postApi.util.upsertQueryData('getCommentAnswers', answersArgs, {
-              pages: [
-                {
-                  items: [optimisticAnswer],
-                  pageSize: COMMENTS_PAGE_SIZE,
-                  totalCount: 1,
-                },
-              ],
-              pageParams: [1],
-            } as unknown as InfiniteData<PaginatedAnswersResponse, number>)
-          )
-          undoAnswer = () => {
-            dispatch(
-              postApi.util.invalidateTags([{ type: 'Comments', id: `${postId}-${commentId}` }])
-            )
-          }
-        }
-
-        const countPatchResult = dispatch(
-          postApi.util.updateQueryData(
-            'getPostComments',
-            { postId, pageSize: COMMENTS_PAGE_SIZE },
-            (draft: InfiniteData<PaginatedCommentsResponse, number>) => {
-              incrementCommentAnswerCount(draft, commentId)
-            }
-          )
-        )
-
-        try {
-          const { data: createdAnswer } = await queryFulfilled
-
-          if (hasCache) {
-            dispatch(
-              postApi.util.updateQueryData(
-                'getCommentAnswers',
-                answersArgs,
-                (draft: InfiniteData<PaginatedAnswersResponse, number>) => {
-                  replaceAnswerInPages(draft, tempId, createdAnswer)
-                }
-              )
-            )
-          } else {
-            dispatch(
-              postApi.util.upsertQueryData('getCommentAnswers', answersArgs, {
-                pages: [
-                  {
-                    items: [createdAnswer],
-                    pageSize: COMMENTS_PAGE_SIZE,
-                    totalCount: 1,
-                  },
-                ],
-                pageParams: [1],
-              } as unknown as InfiniteData<PaginatedAnswersResponse, number>)
-            )
-          }
-        } catch {
-          undoAnswer?.()
-          countPatchResult.undo()
-          showToastAlert({ message: 'Failed to publish answer', type: 'error' })
-        }
-      },
-    }),
   }),
 })
 
@@ -550,10 +269,4 @@ export const {
   useLazyGetPostsByUserQuery,
   useGetPostsQuery,
   useUpdateLikeStatusMutation,
-  useGetPostCommentsInfiniteQuery,
-  useGetCommentAnswersInfiniteQuery,
-  useUpdateCommentLikeStatusMutation,
-  useUpdateAnswerLikeStatusMutation,
-  useCreateCommentMutation,
-  useCreateAnswerMutation,
 } = postApi
