+++ src/shared/types/index.ts
@@ -1,10 +1,9 @@
-// export * from './base/enums'
+export * from './base/enums'
 
-// export * from './auth/dto'
-// export * from './user/models'
-// export * from './messages/message.models'
-// export * from './notifications/notification.models'
-// export * from './posts/models'
-// export * from './comments'
-// export * from './payments/models'
-// export * from './oauth/models'
+export * from './auth/dto'
+export * from './user/models'
+export * from './messages/message.models'
+export * from './notifications/notification.models'
+export * from './posts/models'
+export * from './payments/models'
+export * from './oauth/models'
